fix: add platform-specific imports to resolve dart:html errors on iOS
