+++ lib/utils/stub_utils.dart
@@ -1,0 +1,3 @@
+void addGoogleMapsScript(String apiKey) {
+  // モバイルプラットフォームでは何もしない
+}
